feat: Make input `command` optional, allowing for setup use cases (#7)
This patch marks the `command` input as optional, and skips the step
if it is not specified.

This, combined with the fact that the action installs and adds the CLI
to $PATH, along with properly configured `NUON_CONFIG_FILE` variable,
will allow the users to use the `nuon` CLI directly in this workflows.

Signed-off-by: Prem Saraswat <prem@nuon.co>